% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_load_indicator.R
\name{check_substance_column_names}
\alias{check_substance_column_names}
\title{Check if column names are valid}
\usage{
check_substance_column_names(substances)
}
\arguments{
\item{substances}{Dataframe with substance data}
}
\description{
Check if column names are valid
}
