% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_load_indicator.R
\name{compute_pesticide_load_indicator}
\alias{compute_pesticide_load_indicator}
\title{Compute Pesticide Load Indicator}
\usage{
compute_pesticide_load_indicator(substances, products)
}
\arguments{
\item{substances}{Dataframe with substance data}

\item{products}{Dataframe with products data}
}
\value{
products Dataframe with added columns
}
\description{
Compute Pesticide Load Indicator
}
