% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_load_indicator.R
\name{compute_pesticide_load_indicator_ppdb}
\alias{compute_pesticide_load_indicator_ppdb}
\title{Compute Pesticide Load Indicator with PPDB match}
\usage{
compute_pesticide_load_indicator_ppdb(substances, products, folder)
}
\arguments{
\item{substances}{Dataframe with substance data}

\item{products}{Dataframe with products data}

\item{folder}{folder with exported xlsx files from PPDB}
}
\value{
products Dataframe with added columns
}
\description{
Compute Pesticide Load Indicator with PPDB match
}
