% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_load_indicator.R
\name{compute_pesticide_load_indicator_single}
\alias{compute_pesticide_load_indicator_single}
\title{Compute Pesticide Load Indicator with PPDB match}
\usage{
compute_pesticide_load_indicator_single(
  folder,
  product,
  year,
  formula,
  substances,
  cas_numbers,
  concentrations,
  ...
)
}
\arguments{
\item{folder}{folder with exported xlsx files from PPDB}

\item{product}{product name}

\item{year}{year}

\item{formula}{formulation}

\item{substances}{list or vector of substance names}

\item{cas_numbers}{list or vector of CAS numbers}

\item{concentrations}{list or vector of concentrations}

\item{...}{overrides for default load factors}
}
\value{
products Dataframe with computed pestice loads
}
\description{
Compute Pesticide Load Indicator with PPDB match
}
