% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ppdb.R
\name{match.ppdb}
\alias{match.ppdb}
\title{Extend tables by data from PPDB}
\usage{
match.ppdb(substances, products, folder)
}
\arguments{
\item{substances}{Dataframe with substance data}

\item{products}{Dataframe with products data}

\item{folder}{folder with exported xlsx files from PPDB}
}
\value{
names lists with updated substances and products data frames
}
\description{
Extend tables by data from PPDB
}
